+++ system/lib/libc/musl/src/thread/pthread_mutex_consistent.c
@@ -7,7 +7,7 @@ int pthread_mutex_consistent(pthread_mutex_t *m)
 	int own = old & 0x3fffffff;
 	if (!(m->_m_type & 4) || !own || !(old & 0x40000000))
 		return EINVAL;
-	if (own != CURRENT_THREAD_ID)
+	if (own != __pthread_self()->tid)
 		return EPERM;
 	a_and(&m->_m_lock, ~0x40000000);
 	return 0;